Add table to keep track of progress and bump up RAM for foreman temporarily.
