+++ common/data_refinery_common/models/__init__.py
@@ -1,3 +1,7 @@
+from data_refinery_common.models.command_progress import (
+    CdfCorrectedAccession,
+    SurveyedAccession,
+)
 from data_refinery_common.models.jobs import (
     SurveyJob,
     SurveyJobKeyValue,
@@ -5,7 +9,6 @@ from data_refinery_common.models.jobs import (
     ProcessorJob
 )
 from data_refinery_common.models.organism import Organism
-
 from data_refinery_common.models.models import (
     APIToken,
     ComputationalResult,
@@ -29,5 +32,4 @@ from data_refinery_common.models.models import (
     SampleAnnotation,
     SampleComputedFileAssociation,
     SampleResultAssociation,
-    SurveyedAccession,
 )
